+++ pages/news/[slug].vue
@@ -2,10 +2,17 @@
 import { useRoute, useRouter } from 'vue-router'
 import NewsDetail from '@/components/content/NewsDetail.vue'
 
+const { locale } = useI18n()
 const route = useRoute()
 const router = useRouter()
 const { data: page } = await useAsyncData(`news-${route.params.slug}`, () => {
-  return queryCollection('news').path(`/news/${route.params.slug}`).first()
+  const path = `/news/${locale.value}/${route.params.slug}`
+  return queryCollection('news').path(path).first()
+})
+
+useSeoMeta({
+  title: () => page.value?.title,
+  description: () => page.value?.summary
 })
 </script>